+++ CHANGELOG.md
@@ -6,6 +6,10 @@ A new version of stroom-docs is released for each commit so version tags
 are not recorded in here. Changes are in chronological order with the most
 recent at the top.
 
+* Remove empty `user-guide.md` page
+
+* Add missing links in `SUMMARY.md` to `expressions.md` and `gitbook.md`
+
 * Add `VERSION.md`
 
 * Add `CHANGELOG.md`